// components/Settings/types.ts
/**
 * Enum representing the different sections of the settings dialog
 */
export enum SettingsSection {
  GENERAL = 'GENERAL',
  CHAT_SETTINGS = 'CHAT_SETTINGS',
  AGENT_FEATURES = 'AGENT_FEATURES',
  PRIVACY_CONTROL = 'PRIVACY_CONTROL',
  DATA_MANAGEMENT = 'DATA_MANAGEMENT',
  ACCOUNT = 'ACCOUNT',
  HELP_SUPPORT = 'HELP_SUPPORT',
}